refactor(mcp): split mcp-server god-module
Decompose the ~1430 LOC mcp-server.ts (audit architecture F2 / maintainability
F1) into src/lib/services/mcp/: types, state, target-resolution, dispatch-log,
registry, transport, session, and tools/* (sessions, interaction,
orchestration, ui-writes, introspection, filesystem). mcp-server.ts is now a
111-LOC facade re-exporting initMcpServer and the test hooks, so the public
API and all _*ForTest imports are unchanged. Pure structural change.
