+++ src/lib/services/mcp-server.ts
@@ -18,1363 +18,52 @@
  *
  * We hand-roll JSON-RPC 2.0 rather than pulling in `@modelcontextprotocol/sdk`
  * for two reasons: the SDK targets Node, and the surface area is tiny.
+ *
+ * This module is now a thin facade. The implementation is decomposed under
+ * `./mcp/`:
+ *   - `mcp/types.ts`             — shared types + constants
+ *   - `mcp/state.ts`            — session / connection maps
+ *   - `mcp/target-resolution.ts` — connection-binding → workspace/pane resolution
+ *   - `mcp/dispatch-log.ts`      — observability dispatch log
+ *   - `mcp/registry.ts`          — tool registry
+ *   - `mcp/tools/*`              — individual tool handlers (self-register)
+ *   - `mcp/transport.ts`         — JSON-RPC framing / dispatch
+ *   - `mcp/session.ts`           — wiring / lifecycle (initMcpServer)
+ *
+ * Importing this module registers all tools (via `./mcp/tools`) and re-exports
+ * the public + test surface so existing imports keep resolving unchanged.
  */
-import { get } from "svelte/store";
-import { invoke } from "@tauri-apps/api/core";
-import { listen, emit } from "@tauri-apps/api/event";
+import "./mcp/tools";
+
 import {
-  workspaces,
-  activeWorkspace,
-  activePane,
-} from "../stores/workspace";
-import {
-  getAllPanes,
-  getAllSurfaces,
-  isTerminalSurface,
-  uid,
-  findPaneById as findPaneByIdInWorkspaces,
-  type Pane,
-  type SplitNode,
-  type Surface,
-  type TerminalSurface,
-  type Workspace,
-} from "../types";
-import { findParentSplit } from "../types";
-import { createTerminalSurface, waitForPtyReady } from "../terminal-service";
-import { safeFocus } from "./service-helpers";
-import {
-  registerMcpPty,
-  unregisterMcpPty,
-  getMcpBuffer,
-} from "./mcp-output-buffer";
-import { pushEvent, pollEvents } from "./mcp-event-buffer";
-import {
-  upsertSection,
-  removeSection,
-  type SidebarItem,
-} from "../stores/extension-sidebar";
-import { openPreviewFromContent } from "../../preview";
-import { getMcpSetting } from "../config";
-
-// ---- Types ----
-
-type AgentType = "claude-code" | "codex" | "aider" | "custom";
-type SessionStatus = "starting" | "running" | "idle" | "exited";
-
-interface McpSession {
-  session_id: string;
-  name: string;
-  agent: AgentType;
-  pid: number | undefined;
-  status: SessionStatus;
-  cwd: string;
-  createdAt: string;
-  paneId: string;
-  surfaceId: string;
-  ptyId: number;
-}
-
-interface JsonRpcRequest {
-  jsonrpc: "2.0";
-  id?: number | string | null;
-  method: string;
-  params?: unknown;
-}
-
-interface JsonRpcSuccess {
-  jsonrpc: "2.0";
-  id: number | string | null;
-  result: unknown;
-}
-
-interface JsonRpcError {
-  jsonrpc: "2.0";
-  id: number | string | null;
-  error: { code: number; message: string; data?: unknown };
-}
-
-type JsonRpcResponse = JsonRpcSuccess | JsonRpcError;
-
-/** Per-connection state recorded from the `$/gnar-term/hello` handshake. */
-export interface ConnectionBinding {
-  paneId: string | null;
-  workspaceId: string | null;
-  clientPid: number | null;
-}
-
-export interface ConnectionContext {
-  connectionId: number;
-  binding: ConnectionBinding | null;
-  /** Most-recent pane this connection spawned into. Used as the split host for
-   *  the *next* spawn so rapid-fire `dispatch_tasks` produces a shallow
-   *  right-chain instead of an N-deep left spine around the binding pane —
-   *  the latter exploded `findParentSplit` + DOM render cost into O(N²). */
-  lastSpawnedPaneId: string | null;
-}
-
-/** Sentinel for callers that have no transport context (test code calling
- *  dispatch directly, etc). Tools that require binding will error with the
- *  standard "agent has no pane/workspace context" message — exactly the same
- *  error path real unbound agents hit. */
-const ANONYMOUS_CONTEXT: ConnectionContext = {
-  connectionId: 0,
-  binding: null,
-  lastSpawnedPaneId: null,
-};
-
-// ---- Agent command map ----
-
-const AGENT_COMMANDS: Record<AgentType, string | null> = {
-  "claude-code": "claude",
-  codex: "codex",
-  aider: "aider",
-  custom: null,
-};
-
-const KEY_MAP: Record<string, string> = {
-  "ctrl+c": "\x03",
-  "ctrl+d": "\x04",
-  "ctrl+z": "\x1a",
-  "ctrl+l": "\x0c",
-  "ctrl+a": "\x01",
-  "ctrl+e": "\x05",
-  "ctrl+u": "\x15",
-  "ctrl+k": "\x0b",
-  enter: "\r",
-  tab: "\t",
-  escape: "\x1b",
-  backspace: "\x7f",
-  up: "\x1b[A",
-  down: "\x1b[B",
-  right: "\x1b[C",
-  left: "\x1b[D",
-};
-
-// ---- Session state ----
-
-const sessions = new Map<string, McpSession>();
-const ptyToSession = new Map<number, string>();
-
-/** Per-connection binding state. Keyed by connection_id from the bridge. */
-const connectionContexts = new Map<number, ConnectionContext>();
-
-function newSessionId(): string {
-  return `mcp-${crypto.randomUUID?.() ?? Math.random().toString(36).slice(2)}`;
-}
-
-// ---- Workspace / pane resolution ----
-
-function findPaneById(paneId: string): { workspace: Workspace; pane: Pane } | null {
-  return findPaneByIdInWorkspaces(get(workspaces), paneId);
-}
-
-function findWorkspaceById(workspaceId: string): Workspace | null {
-  return get(workspaces).find((w) => w.id === workspaceId) ?? null;
-}
-
-/** Resolution result returned by `resolveTarget`. `hostPane` is non-null when
- *  the caller expressed a specific host pane to split; null means "place in
- *  the workspace's active pane (or first pane)." */
-interface ResolvedTarget {
-  workspace: Workspace;
-  hostPane: Pane | null;
-  source:
-    | "args-pane"
-    | "args-workspace"
-    | "binding-pane"
-    | "binding-workspace";
-}
-
-interface TargetArgs {
-  workspace_id?: string;
-  pane_id?: string;
-}
-
-/** Resolve which workspace (and optional host pane) a UI-mutating tool should
- *  target. Implements the resolution rules in the MCP spec § Connection binding.
- *  Throws with an actionable error message if no target can be resolved.
- *  Critically: this function NEVER reads `activeWorkspace`. User GUI focus is
- *  not an authoritative routing input. */
-function resolveTarget(args: TargetArgs, ctx: ConnectionContext): ResolvedTarget {
-  // Rule 1: explicit pane_id wins.
-  if (args.pane_id) {
-    const found = findPaneById(args.pane_id);
-    if (!found) {
-      throw new Error(
-        `pane_id "${args.pane_id}" not found (it may have been closed)`,
-      );
-    }
-    return { workspace: found.workspace, hostPane: found.pane, source: "args-pane" };
-  }
-  // Rule 2: explicit workspace_id.
-  if (args.workspace_id) {
-    const ws = findWorkspaceById(args.workspace_id);
-    if (!ws) {
-      throw new Error(`workspace_id "${args.workspace_id}" not found`);
-    }
-    return { workspace: ws, hostPane: null, source: "args-workspace" };
-  }
-  // Rule 3a: chain off the most recent pane this connection spawned, if any.
-  // Keeps the split tree shallow under rapid-fire spawns — see the comment on
-  // `ConnectionContext.lastSpawnedPaneId`.
-  if (ctx.lastSpawnedPaneId) {
-    const found = findPaneById(ctx.lastSpawnedPaneId);
-    if (found) {
-      return {
-        workspace: found.workspace,
-        hostPane: found.pane,
-        source: "binding-pane",
-      };
-    }
-    // Last pane was closed. Fall through.
-  }
-  // Rule 3b: connection-bound pane (re-derive workspace in case pane was moved).
-  const binding = ctx.binding;
-  if (binding?.paneId) {
-    const found = findPaneById(binding.paneId);
-    if (found) {
-      return {
-        workspace: found.workspace,
-        hostPane: found.pane,
-        source: "binding-pane",
-      };
-    }
-    // Bound pane was closed. Fall through to rule 4 (workspace-only binding).
-  }
-  // Rule 4: connection-bound workspace.
-  if (binding?.workspaceId) {
-    const ws = findWorkspaceById(binding.workspaceId);
-    if (ws) {
-      return { workspace: ws, hostPane: null, source: "binding-workspace" };
-    }
-  }
-  // Rule 5: no target. Error loudly — never fall back to active workspace.
-  throw new Error(
-    "agent has no pane/workspace context — pass workspace_id explicitly, or run the agent inside a gnar-term pane",
-  );
-}
-
-/** Pick a host pane to split off when the caller didn't supply one. Prefers
- *  the workspace's active pane; falls back to the first pane in the tree. */
-function pickHostPane(workspace: Workspace): Pane {
-  const all = getAllPanes(workspace.splitRoot);
-  if (workspace.activePaneId) {
-    const active = all.find((p) => p.id === workspace.activePaneId);
-    if (active) return active;
-  }
-  if (all.length > 0) return all[0];
-  // Workspace exists but has no panes (shouldn't happen in practice; create one).
-  const newPane: Pane = { id: uid(), surfaces: [], activeSurfaceId: null };
-  workspace.splitRoot = { type: "pane", pane: newPane };
-  workspaces.update((l) => [...l]);
-  return newPane;
-}
-
-/** Split `hostPane` off into a new sibling pane within its workspace. */
-function splitPaneInWorkspace(
-  workspace: Workspace,
-  hostPane: Pane,
-  direction: "horizontal" | "vertical",
-): Pane {
-  const newPane: Pane = { id: uid(), surfaces: [], activeSurfaceId: null };
-  const newSplit: SplitNode = {
-    type: "split",
-    direction,
-    children: [
-      { type: "pane", pane: hostPane },
-      { type: "pane", pane: newPane },
-    ],
-    ratio: 0.5,
-  };
-  if (
-    workspace.splitRoot.type === "pane" &&
-    workspace.splitRoot.pane.id === hostPane.id
-  ) {
-    workspace.splitRoot = newSplit;
-  } else {
-    const parentInfo = findParentSplit(workspace.splitRoot, hostPane.id);
-    if (parentInfo && parentInfo.parent.type === "split") {
-      parentInfo.parent.children[parentInfo.index] = newSplit;
-    }
-  }
-  workspace.activePaneId = newPane.id;
-  workspaces.update((l) => [...l]);
-  return newPane;
-}
-
-async function waitForPtyId(
-  surface: TerminalSurface,
-  timeoutMs = 5000,
-): Promise<number> {
-  return waitForPtyReady(surface, timeoutMs);
-}
-
-async function getPtyCwd(ptyId: number): Promise<string> {
-  try {
-    return await invoke<string>("get_pty_cwd", { ptyId });
-  } catch {
-    return "";
-  }
-}
-
-function removeSurfaceFromPane(paneId: string, surfaceId: string): void {
-  workspaces.update((list) => {
-    for (const ws of list) {
-      for (const pane of getAllPanes(ws.splitRoot)) {
-        if (pane.id !== paneId) continue;
-        const idx = pane.surfaces.findIndex((s) => s.id === surfaceId);
-        if (idx < 0) continue;
-        const surface = pane.surfaces[idx];
-        if (isTerminalSurface(surface)) {
-          try {
-            surface.terminal.dispose();
-          } catch {
-            /* ignore */
-          }
-        }
-        pane.surfaces.splice(idx, 1);
-        if (pane.surfaces.length === 0) {
-          pane.activeSurfaceId = null;
-        } else {
-          pane.activeSurfaceId =
-            pane.surfaces[Math.min(idx, pane.surfaces.length - 1)].id;
-        }
-      }
-    }
-    return [...list];
-  });
-}
-
-function reapDeadSessions(): void {
-  const aliveSurfaceIds = new Set<string>();
-  for (const ws of get(workspaces)) {
-    for (const surface of getAllSurfaces(ws)) {
-      aliveSurfaceIds.add(surface.id);
-    }
-  }
-  for (const [id, session] of sessions) {
-    if (!aliveSurfaceIds.has(session.surfaceId)) {
-      unregisterMcpPty(session.ptyId);
-      ptyToSession.delete(session.ptyId);
-      sessions.delete(id);
-    }
-  }
-}
-
-// ---- Workspace introspection helpers ----
-
-function describeSurface(s: Surface) {
-  return { id: s.id, kind: s.kind, title: s.title };
-}
-
-function describePane(pane: Pane, workspaceId: string) {
-  const activeSurface = pane.surfaces.find((s) => s.id === pane.activeSurfaceId);
-  let cwd = "";
-  if (activeSurface && isTerminalSurface(activeSurface)) {
-    cwd = activeSurface.cwd ?? "";
-  }
-  return {
-    id: pane.id,
-    workspaceId,
-    cwd,
-    activeSurfaceId: pane.activeSurfaceId,
-    surfaces: pane.surfaces.map(describeSurface),
-  };
-}
-
-// ---- Observability: dispatch log ----
-
-interface DispatchLogEntry {
-  ts: string;
-  connectionId: number;
-  tool: string;
-  binding: ConnectionBinding | null;
-  args: unknown;
-  resolved?: { workspaceId: string; paneId: string | null; source: string };
-  result?: { kind: "ok"; summary: string } | { kind: "error"; message: string };
-}
-
-const DISPATCH_LOG_MAX = 500;
-const dispatchLog: DispatchLogEntry[] = [];
-
-function logDispatch(entry: DispatchLogEntry): void {
-  dispatchLog.push(entry);
-  if (dispatchLog.length > DISPATCH_LOG_MAX) {
-    dispatchLog.shift();
-  }
-  // Echo to console in a structured single line so devtools can grep.
-  const resolved = entry.resolved
-    ? `resolved={ws=${entry.resolved.workspaceId},pane=${entry.resolved.paneId ?? "-"},src=${entry.resolved.source}}`
-    : "resolved=<unresolved>";
-  const bind = entry.binding
-    ? `binding={ws=${entry.binding.workspaceId ?? "null"},pane=${entry.binding.paneId ?? "null"}}`
-    : "binding=<none>";
-  const result = entry.result
-    ? entry.result.kind === "ok"
-      ? `result=ok(${entry.result.summary})`
-      : `result=ERR(${entry.result.message})`
-    : "result=<pending>";
-  // eslint-disable-next-line no-console
-  console.log(
-    `[mcp] conn=#${entry.connectionId} tool=${entry.tool} ${bind} args=${JSON.stringify(entry.args)} ${resolved} ${result}`,
-  );
-}
-
-export function getDispatchLog(): readonly DispatchLogEntry[] {
-  return dispatchLog;
-}
-
-// ---- Tool definitions ----
-
-interface ToolDef {
-  name: string;
-  description: string;
-  inputSchema: Record<string, unknown>;
-  handler: (
-    args: Record<string, unknown>,
-    ctx: ConnectionContext,
-  ) => Promise<unknown> | unknown;
-}
-
-const TOOLS: ToolDef[] = [];
-
-function registerTool(t: ToolDef) {
-  TOOLS.push(t);
-}
-
-// ---- Session management ----
-
-registerTool({
-  name: "spawn_agent",
-  description:
-    "Spawn a new gnar-term pane running an AI coding agent (claude-code, codex, aider) or a custom command. Targets the agent's host workspace by default (per connection binding); pass workspace_id/pane_id to override.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      name: { type: "string" },
-      agent: { type: "string", enum: ["claude-code", "codex", "aider", "custom"] },
-      task: { type: "string" },
-      cwd: { type: "string" },
-      command: { type: "string" },
-      env: { type: "object", additionalProperties: { type: "string" } },
-      cols: { type: "number" },
-      rows: { type: "number" },
-      workspace_id: { type: "string" },
-      pane_id: { type: "string" },
-    },
-    required: ["name", "agent"],
-  },
-  handler: async (args, ctx) => {
-    const p = args as {
-      name: string;
-      agent: AgentType;
-      task?: string;
-      cwd?: string;
-      command?: string;
-      workspace_id?: string;
-      pane_id?: string;
-    };
-    let startupCommand: string | undefined;
-    if (p.agent === "custom") {
-      if (!p.command) throw new Error('agent "custom" requires a command parameter');
-      startupCommand = p.command;
-    } else {
-      const agentCmd = AGENT_COMMANDS[p.agent];
-      if (!agentCmd) throw new Error(`unknown agent: ${p.agent}`);
-      startupCommand = agentCmd;
-    }
-
-    const target = resolveTarget(p, ctx);
-    const hostPane = target.hostPane ?? pickHostPane(target.workspace);
-    const newPane = splitPaneInWorkspace(target.workspace, hostPane, "vertical");
-    ctx.lastSpawnedPaneId = newPane.id;
-
-    const surface = await createTerminalSurface(newPane, p.cwd);
-    surface.title = p.name;
-    surface.startupCommand = startupCommand;
-    newPane.activeSurfaceId = surface.id;
-    workspaces.update((l) => [...l]);
-    safeFocus(surface);
-
-    const ptyId = await waitForPtyId(surface);
-    registerMcpPty(ptyId);
-
-    const cwd = p.cwd || (await getPtyCwd(ptyId));
-    let pid: number | undefined;
-    try {
-      pid = await invoke<number>("get_pty_pid", { ptyId });
-    } catch {
-      pid = undefined;
-    }
-
-    const session: McpSession = {
-      session_id: newSessionId(),
-      name: p.name,
-      agent: p.agent,
-      pid,
-      status: "starting",
-      cwd,
-      createdAt: new Date().toISOString(),
-      paneId: newPane.id,
-      surfaceId: surface.id,
-      ptyId,
-    };
-    sessions.set(session.session_id, session);
-    ptyToSession.set(ptyId, session.session_id);
-    pushEvent({
-      type: "pane.created",
-      paneId: newPane.id,
-      workspaceId: target.workspace.id,
-    });
-    pushEvent({
-      type: "session.statusChanged",
-      sessionId: session.session_id,
-      status: "starting",
-    });
-
-    if (p.task) {
-      setTimeout(() => {
-        invoke("write_pty", { ptyId, data: p.task + "\r" }).catch(() => {});
-      }, 3000);
-    }
-
-    return {
-      session_id: session.session_id,
-      name: session.name,
-      agent: session.agent,
-      pid: session.pid,
-      status: session.status,
-      cwd: session.cwd,
-      pane_id: newPane.id,
-      workspace_id: target.workspace.id,
-    };
-  },
-});
-
-registerTool({
-  name: "list_sessions",
-  description: "List MCP-spawned sessions currently alive in gnar-term.",
-  inputSchema: { type: "object", properties: {} },
-  handler: () => {
-    reapDeadSessions();
-    const list = Array.from(sessions.values()).map((s) => ({
-      session_id: s.session_id,
-      name: s.name,
-      agent: s.agent,
-      pid: s.pid,
-      status: s.status,
-      cwd: s.cwd,
-      createdAt: s.createdAt,
-    }));
-    return { sessions: list };
-  },
-});
-
-registerTool({
-  name: "get_session_info",
-  description: "Return metadata and buffer stats for an MCP session.",
-  inputSchema: {
-    type: "object",
-    properties: { session_id: { type: "string" } },
-    required: ["session_id"],
-  },
-  handler: (args) => {
-    const { session_id } = args as { session_id: string };
-    const session = sessions.get(session_id);
-    if (!session) throw new Error(`session ${session_id} not found`);
-    const buffer = getMcpBuffer(session.ptyId);
-    return {
-      session_id: session.session_id,
-      name: session.name,
-      agent: session.agent,
-      pid: session.pid,
-      status: session.status,
-      cwd: session.cwd,
-      createdAt: session.createdAt,
-      bufferStats: buffer
-        ? { cursor: buffer.getCursor(), lastLine: buffer.getLastLine() }
-        : null,
-    };
-  },
-});
-
-registerTool({
-  name: "kill_session",
-  description: "Kill an MCP session and close its pane.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      session_id: { type: "string" },
-      signal: { type: "string" },
-    },
-    required: ["session_id"],
-  },
-  handler: async (args) => {
-    const { session_id } = args as { session_id: string };
-    const session = sessions.get(session_id);
-    if (!session) throw new Error(`session ${session_id} not found`);
-    try {
-      await invoke("kill_pty", { ptyId: session.ptyId });
-    } catch (err) {
-      console.warn("kill_pty failed:", err);
-    }
-    removeSurfaceFromPane(session.paneId, session.surfaceId);
-    unregisterMcpPty(session.ptyId);
-    ptyToSession.delete(session.ptyId);
-    sessions.delete(session.session_id);
-    pushEvent({
-      type: "session.statusChanged",
-      sessionId: session.session_id,
-      status: "exited",
-    });
-    return { ok: true };
-  },
-});
-
-// ---- Interaction ----
-
-registerTool({
-  name: "send_prompt",
-  description: "Send text to an MCP session's PTY. Appends Enter unless press_enter is false.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      session_id: { type: "string" },
-      text: { type: "string" },
-      press_enter: { type: "boolean" },
-    },
-    required: ["session_id", "text"],
-  },
-  handler: async (args) => {
-    const p = args as { session_id: string; text: string; press_enter?: boolean };
-    const session = sessions.get(p.session_id);
-    if (!session) throw new Error(`session ${p.session_id} not found`);
-    const data = p.text + (p.press_enter === false ? "" : "\r");
-    await invoke("write_pty", { ptyId: session.ptyId, data });
-    return { ok: true };
-  },
-});
-
-registerTool({
-  name: "send_keys",
-  description: "Send a named keystroke (ctrl+c, enter, escape, arrows, etc.) to an MCP session.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      session_id: { type: "string" },
-      keys: { type: "string" },
-    },
-    required: ["session_id", "keys"],
-  },
-  handler: async (args) => {
-    const p = args as { session_id: string; keys: string };
-    const session = sessions.get(p.session_id);
-    if (!session) throw new Error(`session ${p.session_id} not found`);
-    const sequence = KEY_MAP[p.keys.toLowerCase()];
-    if (!sequence) {
-      throw new Error(
-        `unknown key "${p.keys}". Available: ${Object.keys(KEY_MAP).join(", ")}`,
-      );
-    }
-    await invoke("write_pty", { ptyId: session.ptyId, data: sequence });
-    return { ok: true };
-  },
-});
-
-registerTool({
-  name: "read_output",
-  description: "Read terminal output from an MCP session. Supports cursor-based polling and ANSI stripping.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      session_id: { type: "string" },
-      lines: { type: "number" },
-      cursor: { type: "number" },
-      strip_ansi: { type: "boolean" },
-    },
-    required: ["session_id"],
-  },
-  handler: (args) => {
-    const p = args as {
-      session_id: string;
-      lines?: number;
-      cursor?: number;
-      strip_ansi?: boolean;
-    };
-    const session = sessions.get(p.session_id);
-    if (!session) throw new Error(`session ${p.session_id} not found`);
-    const buffer = getMcpBuffer(session.ptyId);
-    if (!buffer) {
-      return {
-        output: "",
-        cursor: 0,
-        total_lines: 0,
-        session_status: session.status,
-      };
-    }
-    const result = buffer.read({
-      lines: p.lines,
-      cursor: p.cursor,
-      strip_ansi: p.strip_ansi,
-    });
-    return { ...result, session_status: session.status };
-  },
-});
-
-// ---- Orchestration ----
-
-registerTool({
-  name: "dispatch_tasks",
-  description:
-    "Spawn multiple agent sessions in parallel. Each task resolves its target independently — pass workspace_id/pane_id per task to override the connection binding.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      tasks: {
-        type: "array",
-        items: {
-          type: "object",
-          properties: {
-            name: { type: "string" },
-            agent: { type: "string" },
-            task: { type: "string" },
-            cwd: { type: "string" },
-            command: { type: "string" },
-            workspace_id: { type: "string" },
-            pane_id: { type: "string" },
-          },
-          required: ["name", "agent", "task"],
-        },
-      },
-    },
-    required: ["tasks"],
-  },
-  handler: async (args, ctx) => {
-    const { tasks } = args as {
-      tasks: Array<{
-        name: string;
-        agent: AgentType;
-        task: string;
-        cwd?: string;
-        command?: string;
-        workspace_id?: string;
-        pane_id?: string;
-      }>;
-    };
-    const results: Array<{
-      session_id: string;
-      name: string;
-      agent: AgentType;
-      pid: number | undefined;
-      pane_id?: string;
-      workspace_id?: string;
-      error?: string;
-    }> = [];
-    const spawnTool = TOOLS.find((t) => t.name === "spawn_agent")!;
-    for (const taskDef of tasks) {
-      try {
-        const resp = (await spawnTool.handler(
-          {
-            name: taskDef.name,
-            agent: taskDef.agent,
-            task: taskDef.task,
-            cwd: taskDef.cwd,
-            command: taskDef.command,
-            workspace_id: taskDef.workspace_id,
-            pane_id: taskDef.pane_id,
-          },
-          ctx,
-        )) as {
-          session_id: string;
-          name: string;
-          agent: AgentType;
-          pid: number | undefined;
-          pane_id: string;
-          workspace_id: string;
-        };
-        results.push({
-          session_id: resp.session_id,
-          name: resp.name,
-          agent: resp.agent,
-          pid: resp.pid,
-          pane_id: resp.pane_id,
-          workspace_id: resp.workspace_id,
-        });
-      } catch (err) {
-        results.push({
-          session_id: "",
-          name: taskDef.name,
-          agent: taskDef.agent,
-          pid: undefined,
-          error: err instanceof Error ? err.message : String(err),
-        });
-      }
-    }
-    return {
-      dispatched: results.filter((r) => !r.error).length,
-      failed: results.filter((r) => r.error).length,
-      sessions: results,
-    };
-  },
-});
-
-// ---- UI writes ----
-
-registerTool({
-  name: "render_sidebar",
-  description:
-    "Declare or replace an extension sidebar section in the resolved workspace. Sections are workspace-scoped: invisible from other workspaces.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      section_id: { type: "string" },
-      title: { type: "string" },
-      items: { type: "array" },
-      workspace_id: { type: "string" },
-    },
-    required: ["section_id", "title", "items"],
-  },
-  handler: (args, ctx) => {
-    const p = args as {
-      section_id: string;
-      title: string;
-      items: SidebarItem[];
-      workspace_id?: string;
-    };
-    const target = resolveTarget({ workspace_id: p.workspace_id }, ctx);
-    upsertSection({
-      sectionId: p.section_id,
-      title: p.title,
-      items: p.items ?? [],
-      workspaceId: target.workspace.id,
-    });
-    return { ok: true, workspace_id: target.workspace.id };
-  },
-});
-
-registerTool({
-  name: "remove_sidebar_section",
-  description:
-    "Remove an extension-declared sidebar section from the resolved workspace. Safe for non-existent IDs.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      section_id: { type: "string" },
-      workspace_id: { type: "string" },
-    },
-    required: ["section_id"],
-  },
-  handler: (args, ctx) => {
-    const p = args as {
-      section_id: string;
-      workspace_id?: string;
-    };
-    const target = resolveTarget({ workspace_id: p.workspace_id }, ctx);
-    removeSection(target.workspace.id, p.section_id);
-    return { ok: true, workspace_id: target.workspace.id };
-  },
-});
-
-registerTool({
-  name: "create_preview",
-  description:
-    "Open a preview surface with markdown/text/code content. Targets the agent's host workspace by default; pass workspace_id/pane_id to override.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      content: { type: "string" },
-      format: { type: "string", enum: ["markdown", "text", "code"] },
-      language: { type: "string" },
-      title: { type: "string" },
-      placement: {
-        type: "string",
-        enum: ["split-right", "split-down", "new-tab", "current-pane"],
-      },
-      workspace_id: { type: "string" },
-      pane_id: { type: "string" },
-    },
-    required: ["content", "format"],
-  },
-  handler: (args, ctx) => {
-    const p = args as {
-      content: string;
-      format: "markdown" | "text" | "code";
-      language?: string;
-      title?: string;
-      placement?: "split-right" | "split-down" | "new-tab" | "current-pane";
-      workspace_id?: string;
-      pane_id?: string;
-    };
-    const target = resolveTarget(p, ctx);
-    const title = p.title ?? "Preview";
-    let rendered: string;
-    if (p.format === "markdown") rendered = p.content;
-    else if (p.format === "code")
-      rendered = "```" + (p.language ?? "") + "\n" + p.content + "\n```";
-    else rendered = "```\n" + p.content + "\n```";
-    const previewSurface = openPreviewFromContent(rendered, title);
-
-    const placement = p.placement ?? "split-right";
-    const hostPane = target.hostPane ?? pickHostPane(target.workspace);
-    let targetPane: Pane;
-    if (placement === "split-right") {
-      targetPane = splitPaneInWorkspace(target.workspace, hostPane, "horizontal");
-      ctx.lastSpawnedPaneId = targetPane.id;
-    } else if (placement === "split-down") {
-      targetPane = splitPaneInWorkspace(target.workspace, hostPane, "vertical");
-      ctx.lastSpawnedPaneId = targetPane.id;
-    } else {
-      // new-tab and current-pane drop into the host pane's surface list.
-      targetPane = hostPane;
-    }
-
-    const surface = {
-      kind: "preview" as const,
-      id: previewSurface.id,
-      filePath: previewSurface.filePath,
-      title: previewSurface.title,
-      element: previewSurface.element,
-      watchId: previewSurface.watchId,
-      hasUnread: false,
-    };
-    targetPane.surfaces.push(surface);
-    targetPane.activeSurfaceId = surface.id;
-    workspaces.update((l) => [...l]);
-
-    return {
-      preview_id: surface.id,
-      pane_id: targetPane.id,
-      workspace_id: target.workspace.id,
-    };
-  },
-});
-
-// ---- Agent introspection ----
-
-registerTool({
-  name: "get_agent_context",
-  description:
-    "Return the connection's bound {pane_id, workspace_id, client_pid} from the $/gnar-term/hello handshake. All fields are null when the agent is unbound (was not run inside a gnar-term pane). Agents should call this once on startup to learn their context.",
-  inputSchema: { type: "object", properties: {} },
-  handler: (_args, ctx) => {
-    const b = ctx.binding;
-    return {
-      pane_id: b?.paneId ?? null,
-      workspace_id: b?.workspaceId ?? null,
-      client_pid: b?.clientPid ?? null,
-    };
-  },
-});
-
-// ---- UI introspection (observers; report user GUI focus) ----
-
-registerTool({
-  name: "get_active_workspace",
-  description:
-    "Return the workspace the user is currently focused on. Reports user GUI focus, NOT the agent's binding — agents should use get_agent_context for routing. Fields are null when no workspace is open.",
-  inputSchema: { type: "object", properties: {} },
-  handler: () => {
-    const ws = get(activeWorkspace);
-    return {
-      id: ws?.id ?? null,
-      name: ws?.name ?? null,
-      activePaneId: ws?.activePaneId ?? null,
-    };
-  },
-});
-
-registerTool({
-  name: "list_workspaces",
-  description: "List all open workspaces.",
-  inputSchema: { type: "object", properties: {} },
-  handler: () => {
-    const list = get(workspaces).map((ws) => ({
-      id: ws.id,
-      name: ws.name,
-      activePaneId: ws.activePaneId,
-    }));
-    return { workspaces: list };
-  },
-});
-
-registerTool({
-  name: "get_active_pane",
-  description:
-    "Return the user-focused pane and its surfaces. Reports user GUI focus, NOT the agent's binding. `pane` is null when no pane is focused.",
-  inputSchema: { type: "object", properties: {} },
-  handler: () => {
-    const ws = get(activeWorkspace);
-    const pane = get(activePane);
-    if (!ws || !pane) return { pane: null };
-    return { pane: describePane(pane, ws.id) };
-  },
-});
-
-registerTool({
-  name: "list_panes",
-  description: "List panes in a workspace (defaults to the active workspace).",
-  inputSchema: {
-    type: "object",
-    properties: { workspace_id: { type: "string" } },
-  },
-  handler: (args) => {
-    const p = args as { workspace_id?: string };
-    const target = p.workspace_id
-      ? get(workspaces).find((w) => w.id === p.workspace_id)
-      : get(activeWorkspace);
-    if (!target) return { panes: [] };
-    const list = getAllPanes(target.splitRoot).map((pane) => describePane(pane, target.id));
-    return { panes: list };
-  },
-});
-
-// ---- Lifecycle events ----
-
-registerTool({
-  name: "poll_events",
-  description: "Poll the 500-entry lifecycle event ring buffer.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      cursor: { type: "number" },
-      max: { type: "number" },
-    },
-  },
-  handler: (args) => {
-    const p = args as { cursor?: number; max?: number };
-    return pollEvents({ cursor: p.cursor, max: p.max });
-  },
-});
-
-// ---- Filesystem ----
-
-registerTool({
-  name: "list_dir",
-  description: "List a directory. Returns entries with name, path, is_dir, size.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      path: { type: "string" },
-      include_hidden: { type: "boolean" },
-    },
-    required: ["path"],
-  },
-  handler: async (args) => {
-    const p = args as { path: string; include_hidden?: boolean };
-    const entries = await invoke<
-      Array<{ name: string; path: string; is_dir: boolean; size: number }>
-    >("mcp_list_dir", { path: p.path, includeHidden: p.include_hidden });
-    return { entries };
-  },
-});
-
-registerTool({
-  name: "read_file",
-  description: "Read a UTF-8 file and return its contents. Non-UTF-8 content is an error.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      path: { type: "string" },
-      max_bytes: { type: "number" },
-    },
-    required: ["path"],
-  },
-  handler: async (args) => {
-    const p = args as { path: string; max_bytes?: number };
-    const content = await invoke<string>("read_file", { path: p.path });
-    if (p.max_bytes && content.length > p.max_bytes) {
-      return { content: content.slice(0, p.max_bytes), truncated: true };
-    }
-    return { content, truncated: false };
-  },
-});
-
-registerTool({
-  name: "file_exists",
-  description: "Check whether a path exists. Returns exists and (if it does) is_dir.",
-  inputSchema: {
-    type: "object",
-    properties: { path: { type: "string" } },
-    required: ["path"],
-  },
-  handler: async (args) => {
-    const p = args as { path: string };
-    const [exists, isDir] = await invoke<[boolean, boolean]>("mcp_file_info", {
-      path: p.path,
-    });
-    return exists ? { exists: true, is_dir: isDir } : { exists: false };
-  },
-});
-
-// ---- JSON-RPC dispatch ----
-
-const PROTOCOL_VERSION = "2025-11-25";
-const SERVER_INFO = { name: "gnar-term", version: "0.3.1" };
-
-/** Tools that mutate UI surfaces — these are the ones that get logged to
- *  dispatchLog with binding/resolution metadata. */
-const UI_MUTATING_TOOLS = new Set([
-  "spawn_agent",
-  "dispatch_tasks",
-  "create_preview",
-  "render_sidebar",
-  "remove_sidebar_section",
-]);
-
-export async function dispatch(
-  req: JsonRpcRequest,
-  ctx: ConnectionContext = ANONYMOUS_CONTEXT,
-): Promise<JsonRpcResponse | null> {
-  const id = req.id ?? null;
-  const isNotification = req.id === undefined || req.id === null;
-
-  try {
-    switch (req.method) {
-      case "initialize": {
-        const result = {
-          protocolVersion: PROTOCOL_VERSION,
-          capabilities: { tools: { listChanged: false } },
-          serverInfo: SERVER_INFO,
-        };
-        return isNotification ? null : { jsonrpc: "2.0", id, result };
-      }
-      case "$/gnar-term/hello": {
-        // Connection-binding handshake. Always a notification (no id).
-        const params = (req.params ?? {}) as {
-          pane_id?: string | null;
-          workspace_id?: string | null;
-          client_pid?: number | null;
-        };
-        ctx.binding = {
-          paneId: params.pane_id ?? null,
-          workspaceId: params.workspace_id ?? null,
-          clientPid: params.client_pid ?? null,
-        };
-        return null;
-      }
-      case "notifications/initialized":
-      case "initialized":
-        return null;
-      case "tools/list": {
-        const tools = TOOLS.map((t) => ({
-          name: t.name,
-          description: t.description,
-          inputSchema: t.inputSchema,
-        }));
-        return isNotification ? null : { jsonrpc: "2.0", id, result: { tools } };
-      }
-      case "tools/call": {
-        const params = (req.params ?? {}) as {
-          name: string;
-          arguments?: Record<string, unknown>;
-        };
-        const tool = TOOLS.find((t) => t.name === params.name);
-        if (!tool) {
-          if (isNotification) return null;
-          return {
-            jsonrpc: "2.0",
-            id,
-            error: { code: -32601, message: `unknown tool: ${params.name}` },
-          };
-        }
-        const args = params.arguments ?? {};
-        const shouldLog = UI_MUTATING_TOOLS.has(tool.name);
-        const logEntry: DispatchLogEntry | null = shouldLog
-          ? {
-              ts: new Date().toISOString(),
-              connectionId: ctx.connectionId,
-              tool: tool.name,
-              binding: ctx.binding,
-              args,
-            }
-          : null;
-        try {
-          const value = await tool.handler(args, ctx);
-          if (logEntry) {
-            // Try to surface the resolution metadata when the result includes it.
-            const v = value as
-              | { workspace_id?: string; pane_id?: string }
-              | undefined
-              | null;
-            if (v && v.workspace_id) {
-              logEntry.resolved = {
-                workspaceId: v.workspace_id,
-                paneId: v.pane_id ?? null,
-                source: "tool-result",
-              };
-            }
-            logEntry.result = { kind: "ok", summary: JSON.stringify(value).slice(0, 120) };
-            logDispatch(logEntry);
-          }
-          if (isNotification) return null;
-          // MCP spec requires `structuredContent` to be a JSON object (record).
-          // Arrays, null, and primitives are rejected by client-side Zod
-          // validators, so only attach structuredContent when the tool returned
-          // a plain object. Text `content` still carries the full JSON payload.
-          const isRecord =
-            value !== null && typeof value === "object" && !Array.isArray(value);
-          return {
-            jsonrpc: "2.0",
-            id,
-            result: {
-              content: [{ type: "text", text: JSON.stringify(value) }],
-              ...(isRecord ? { structuredContent: value } : {}),
-            },
-          };
-        } catch (err) {
-          if (logEntry) {
-            logEntry.result = {
-              kind: "error",
-              message: err instanceof Error ? err.message : String(err),
-            };
-            logDispatch(logEntry);
-          }
-          throw err;
-        }
-      }
-      case "ping":
-        return isNotification ? null : { jsonrpc: "2.0", id, result: {} };
-      default:
-        if (isNotification) return null;
-        return {
-          jsonrpc: "2.0",
-          id,
-          error: { code: -32601, message: `method not found: ${req.method}` },
-        };
-    }
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    if (isNotification) {
-      console.warn("[mcp] notification handler threw:", message);
-      return null;
-    }
-    return {
-      jsonrpc: "2.0",
-      id,
-      error: { code: -32000, message },
-    };
-  }
-}
-
-// ---- Wiring ----
-
-let initialized = false;
-
-interface RequestEnvelope {
-  connection_id: number;
-  payload: string;
-}
-
-interface ConnectionClosedEnvelope {
-  connection_id: number;
-}
-
-function getOrCreateContext(connectionId: number): ConnectionContext {
-  let ctx = connectionContexts.get(connectionId);
-  if (!ctx) {
-    ctx = { connectionId, binding: null, lastSpawnedPaneId: null };
-    connectionContexts.set(connectionId, ctx);
-  }
-  return ctx;
-}
-
-/** Initialize the MCP server. Honors the `mcp` config setting. */
-export async function initMcpServer(): Promise<void> {
-  if (initialized) return;
-  initialized = true;
-
-  const setting = getMcpSetting();
-  if (setting === "off") return;
-
-  // Bridge pty-exit to session status updates.
-  await listen<{ pty_id: number }>("pty-exit", (event) => {
-    const sessionId = ptyToSession.get(event.payload.pty_id);
-    if (!sessionId) return;
-    const session = sessions.get(sessionId);
-    if (session) {
-      session.status = "exited";
-      pushEvent({
-        type: "session.statusChanged",
-        sessionId,
-        status: "exited",
-      });
-    }
-    unregisterMcpPty(event.payload.pty_id);
-    ptyToSession.delete(event.payload.pty_id);
-  });
-
-  await listen<string>("mcp-request", async (event) => {
-    let envelope: RequestEnvelope;
-    try {
-      envelope = JSON.parse(event.payload) as RequestEnvelope;
-    } catch {
-      console.warn("[mcp] malformed request envelope:", event.payload);
-      return;
-    }
-    if (
-      typeof envelope.connection_id !== "number" ||
-      typeof envelope.payload !== "string"
-    ) {
-      console.warn("[mcp] invalid request envelope shape:", envelope);
-      return;
-    }
-    let req: JsonRpcRequest;
-    try {
-      req = JSON.parse(envelope.payload) as JsonRpcRequest;
-    } catch {
-      console.warn("[mcp] malformed request payload:", envelope.payload);
-      return;
-    }
-    const ctx = getOrCreateContext(envelope.connection_id);
-    const resp = await dispatch(req, ctx);
-    if (resp) {
-      const out = {
-        connection_id: envelope.connection_id,
-        payload: JSON.stringify(resp),
-      };
-      await emit("mcp-response", JSON.stringify(out));
-    }
-  });
-
-  await listen<string>("mcp-connection-closed", (event) => {
-    let envelope: ConnectionClosedEnvelope;
-    try {
-      envelope = JSON.parse(event.payload) as ConnectionClosedEnvelope;
-    } catch {
-      console.warn("[mcp] malformed close envelope:", event.payload);
-      return;
-    }
-    connectionContexts.delete(envelope.connection_id);
-  });
-
-  // Lifecycle events for user GUI focus changes (observers).
-  let lastWorkspaceId: string | null = null;
-  let lastPaneId: string | null = null;
-  activeWorkspace.subscribe((ws) => {
-    if (!ws) return;
-    if (ws.id !== lastWorkspaceId) {
-      lastWorkspaceId = ws.id;
-      pushEvent({ type: "workspace.changed", workspaceId: ws.id });
-    }
-  });
-  activePane.subscribe((pane) => {
-    if (!pane) return;
-    if (pane.id !== lastPaneId) {
-      lastPaneId = pane.id;
-      const ws = get(activeWorkspace);
-      pushEvent({
-        type: "pane.focused",
-        paneId: pane.id,
-        workspaceId: ws?.id ?? "",
-      });
-    }
-  });
-}
+  sessions,
+  ptyToSession,
+  connectionContexts,
+} from "./mcp/state";
+import { resolveTarget } from "./mcp/target-resolution";
+import { getTools } from "./mcp/registry";
+import { resetDispatchLogForTest } from "./mcp/dispatch-log";
+import { resetInitializedForTest } from "./mcp/session";
+import type {
+  ConnectionBinding,
+  ConnectionContext,
+  McpSession,
+  ResolvedTarget,
+  TargetArgs,
+  ToolDef,
+} from "./mcp/types";
+
+// ---- Public API ----
+
+export { dispatch } from "./mcp/transport";
+export { initMcpServer } from "./mcp/session";
+export { getDispatchLog } from "./mcp/dispatch-log";
+export type { ConnectionBinding, ConnectionContext } from "./mcp/types";
 
 // ---- Test hooks ----
 
 export function _getToolsForTest(): ToolDef[] {
-  return TOOLS;
+  return getTools();
 }
 
 export function _getSessionsForTest(): Map<string, McpSession> {
@@ -1417,6 +106,6 @@ export function _resetMcpServerForTest(): void {
   sessions.clear();
   ptyToSession.clear();
   connectionContexts.clear();
-  dispatchLog.length = 0;
-  initialized = false;
+  resetDispatchLogForTest();
+  resetInitializedForTest();
 }
